+++ device-sdk/src/main/java/com/maxmind/device/model/FontInfo.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
  * Font profile information based on available system fonts.
  */
 @Serializable
-public data class FontInfo(
+internal data class FontInfo(
     @SerialName("available_fonts")
     val availableFonts: List<String> = emptyList(),
 )